requirements.txt
telethon
python-dotenv
asyncio
fastapi
uvicorn[standard]

google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib